+++ libc/dup.c
@@ -4,7 +4,7 @@
 #include<errno.h>
 int dup(int oldfd){
     uint64_t returnvalue = syscall_1(SYS_dup,oldfd);
-    if(returnvalue < 0){
+    if(((int)returnvalue) < 0){
         errno = -(returnvalue);
         return -1;
     }
