+++ libc/dup2.c
@@ -5,7 +5,7 @@
 int dup2(int oldfd, int newfd)
 {
     uint64_t returnvalue = syscall_2((uint64_t)SYS_dup2,(uint64_t)oldfd,(uint64_t)newfd);
-    if(returnvalue < 0){
+    if(((int)returnvalue) < 0){
         errno = -(returnvalue);
         return -1;
     }
