+++ libc/fork.c
@@ -5,7 +5,7 @@
 pid_t fork(void)
 {
     pid_t returnvalue = syscall_0((uint64_t) SYS_fork);
-    if(returnvalue < 0){
+    if(((int)returnvalue) < 0){
         errno = -(returnvalue);
         return -1;
     }
